feat(serve-static): refactoring
